move find in page to ts
