--- packages/compass-find-in-page/.browserslistrc
@@ -1,1 +1,0 @@
-Electron >= 6
