+++ packages/compass-find-in-page/.depcheckrc
@@ -1,16 +1,11 @@
-ignores: [
-  "@babel/cli",
-  "@babel/register",
-  "core-js",
-  "electron-rebuild",
-  "font-awesome",
-  "hadron-react-components",
-  "karma-electron",
-  "karma-sourcemap-loader",
-  "karma-webpack",
-  "mongodb-connection-model",
-  "resolve",
-  "webpack-bundle-analyzer",
-  "webpack-cli",
-  "mongodb-compass"
-]
+ignores:
+  - '@mongodb-js/prettier-config-compass'
+  - '@mongodb-js/tsconfig-compass'
+  - '@types/chai'
+  - '@types/sinon-chai'
+  - 'sinon'
+  - '@types/chai-dom'
+  - '@types/react'
+  - '@types/react-dom'
+  # Will come in handy when we will start to add tests
+  - '@testing-library/user-event'
