+++ packages/compass-find-in-page/.eslintignore
@@ -1,0 +1,2 @@
+.nyc-output
+dist
