--- packages/compass-find-in-page/.gitignore
@@ -1,11 +1,0 @@
-.DS_Store
-npm-debug.log
-node_modules/
-lib/
-lib-cov
-coverage
-.lock-wscript
-.idea/
-*.iml
-.nvmrc
-.nyc_output
