--- packages/compass-find-in-page/.npmignore
@@ -1,17 +1,0 @@
-.babelrc
-.eslintrc
-.editorconfig
-.nycrc
-.prettierrc
-.travis.yml
-config/
-coverage/
-electron/
-karma.conf.js
-mocha-webpack.opts
-scripts/
-test/
-src/
-.coverage/
-.nyc_output/
-lib/index.html
