--- packages/compass-find-in-page/.nycrc
@@ -1,13 +1,0 @@
-{
-  "exclude": [
-    "config",
-    ".tmp",
-    "test"
-  ],
-  "reporter": [
-    "text",
-    "html"
-  ],
-  "sourceMap": false,
-  "instrument": false
-}
