+++ packages/compass-find-in-page/.prettierignore
@@ -1,0 +1,3 @@
+.nyc_output
+dist
+coverage
