--- packages/compass-find-in-page/.prettierrc
@@ -1,3 +1,0 @@
-{
-  "singleQuote": true
-}
