+++ packages/compass-find-in-page/src/components/compass-find-in-page.spec.ts
