--- packages/compass-find-in-page/src/components/compass-find-in-page/index.js
@@ -1,4 +1,0 @@
-import CompassFindInPage from './compass-find-in-page';
-
-export default CompassFindInPage;
-export { CompassFindInPage };
