--- packages/compass-find-in-page/src/components/find-in-page-input/index.js
@@ -1,2 +1,0 @@
-import FindInPageInput from './find-in-page-input';
-export default FindInPageInput;
