+++ packages/compass-find-in-page/src/index.spec.ts
@@ -1,0 +1,10 @@
+import { expect } from 'chai';
+import CompassPlugin from './index';
+
+describe('Compass Plugin', function () {
+  it('exports activate, deactivate, and metadata', function () {
+    expect(CompassPlugin).to.have.property('activate');
+    expect(CompassPlugin).to.have.property('deactivate');
+    expect(CompassPlugin).to.have.property('metadata');
+  });
+});
