+++ packages/compass-find-in-page/src/stores/index.ts
