+++ packages/compass-find-in-page/src/stores/store.spec.ts
